+++ lib/cocoapods/user_interface/error_report.rb
@@ -112,9 +112,9 @@ EOS
 ```
 EOS
         end
-        
+
         def search_for_exceptions(exception)
-          inspector = GhInspector::Inspector.new "cocoapods", "cocoapods"
+          inspector = GhInspector::Inspector.new 'cocoapods', 'cocoapods'
           message_delegate = UserInterface::InspectorReporter.new
           inspector.search_exception exception, message_delegate
         end