+++ lib/cocoapods/user_interface/inspector_reporter.rb
@@ -6,12 +6,31 @@ module Pod
     #
     class InspectorReporter
       # Called just as the investigation has begun.
-      def inspector_started_query(query, inspector)
+      # Lets the user know that it's looking for an issue.
+      #
+      # @param [query] String unused
+      #
+      # @param [GhInspector::Inspector] inspector
+      #        The current inspector
+      #
+      # @return [void]
+      #
+      def inspector_started_query(_, inspector)
         UI.puts "Looking for related issues on #{inspector.repo_owner}/#{inspector.repo_name}..."
       end
 
-      # Called once the inspector has recieved a report with more than one issue.
-      def inspector_successfully_recieved_report(report, inspector)
+      # Called once the inspector has recieved a report with more than one issue,
+      # showing the top 3 issues, and offering a link to see more.
+      #
+      # @param [GhInspector::InspectionReport] report
+      #        Report a list of the issues
+      #
+      # @param [GhInspector::Inspector] inspector
+      #        The current inspector
+      #
+      # @return [void]
+      #
+      def inspector_successfully_recieved_report(report, _)
         report.issues[0..2].each { |issue| print_issue_full(issue) }
 
         if report.issues.count > 3
@@ -21,15 +40,36 @@ module Pod
       end
 
       # Called once the report has been recieved, but when there are no issues found.
-      def inspector_recieved_empty_report(report, inspector)
-        UI.puts "Found no similar issues. To create a new issue, please visit:"
+      #
+      # @param [GhInspector::InspectionReport] report
+      #        An empty report
+      #
+      # @param [GhInspector::Inspector] inspector
+      #        The current inspector
+      #
+      # @return [void]
+      #
+      def inspector_recieved_empty_report(_, inspector)
+        UI.puts 'Found no similar issues. To create a new issue, please visit:'
         UI.puts "https://github.com/#{inspector.repo_owner}/#{inspector.repo_name}/issues/new"
       end
 
       # Called when there have been networking issues in creating the report.
+      #
+      # @param [Error] error
+      #        The error returned during networking
+      #
+      # @param [String] query
+      #        The original search query
+      #
+      # @param [GhInspector::Inspector] inspector
+      #        The current inspector
+      #
+      # @return [void]
+      #
       def inspector_could_not_create_report(error, query, inspector)
         safe_query = URI.escape query
-        UI.puts "Could not access the GitHub API, you may have better luck via the website."
+        UI.puts 'Could not access the GitHub API, you may have better luck via the website.'
         UI.puts "https://github.com/#{inspector.repo_owner}/#{inspector.repo_name}/search?q=#{safe_query}&type=Issues&utf8=✓"
         UI.puts "Error: #{error.name}"
       end
@@ -41,9 +81,9 @@ module Pod
         UI.puts " - #{issue.title}"
         UI.puts "   #{safe_url} [#{issue.state}] [#{issue.comments} comment#{issue.comments == 1 ? '' : 's'}]"
         UI.puts "   #{pretty_date(issue.updated_at)}"
-        UI.puts ""
+        UI.puts ''
       end
-      
+
       # Taken from http://stackoverflow.com/questions/195740/how-do-you-do-relative-time-in-rails
       def pretty_date(date_string)
         date = Time.parse(date_string)
@@ -61,10 +101,9 @@ module Pod
         when 172_001..518_400 then ((a + 800) / (60 * 60 * 24)).to_i.to_s + ' days ago'
         when 518_400..1_036_800 then 'a week ago'
         when 1_036_801..4_147_204 then ((a + 180_000) / (60 * 60 * 24 * 7)).to_i.to_s + ' weeks ago'
-        else date.strftime("%d %b %Y")
+        else date.strftime('%d %b %Y')
         end
       end
     end
   end
 end
-