+++ spec/unit/user_interface/error_report_spec.rb
@@ -110,16 +110,16 @@ EOS
         message = remove_color(message)
         message.should == '[!] at -'
       end
-    
-      it "handles inspector_successfully_recieved_report" do
+
+      it 'handles inspector_successfully_recieved_report' do
         time = Time.new(2016, 5, 13)
         Time.stubs(:now).returns(time)
-        
+
         url = 'https://api.github.com/search/issues?q=Testing+repo:cocoapods/cocoapods'
-        fixture_json_text = File.read SpecHelper.fixture("github_search_response.json")
+        fixture_json_text = File.read SpecHelper.fixture('github_search_response.json')
         GhInspector::Sidekick.any_instance.expects(:get_api_results).with(url).returns(JSON.parse(fixture_json_text))
-        
-        error = NameError.new("Testing", "orta")
+
+        error = NameError.new('Testing', 'orta')
         @report.search_for_exceptions error
         result = <<-EOS
 Looking for related issues on cocoapods/cocoapods...
@@ -140,7 +140,6 @@ https://github.com/cocoapods/cocoapods/search?q=Testing&type=Issues&utf8=✓
 EOS
         UI.output.should == result
       end
-      
     end
   end
 end